tests/AttrExodus.html

<!DOCTYPE html>
<title>AttrExodus</title>
<script src="testharness.js"></script>
<div id="log"></div>
<script>
test(function() {
  document.body.setAttribute("abc", "pass")
  var attr = document.body.attributes[0]
  assert_true(attr instanceof Attr)
  assert_false(attr instanceof Node)
  assert_throws(new TypeError(), function() { attr.appendChild(document.createTextNode("fail")) })
  assert_throws(new TypeError(), function() { attr.appendChild(null) })
  assert_equals(attr.value, "pass")
  assert_false("childNodes" in attr)
});
</script>
